enabling warnings in DS-Game-server x86/x64
- Enable project warnnings /w3
- _CRT_SECURE_NO_WARNINGS to avoid MS secure functions that would break
cross-platform
- Add _CRT_NONSTDC_NO_DEPRECATE/_WINSOCK_DEPRECATED_NO_WARNINGS to
silence POSIX/deprecated winsock methods
- most issues are type conversions
- some involved type changes
- some involved loop conversion to ranged-based
- std::any/std::any_cast to handle void* for CTaskMgr

+++ src/common/showmsg.cpp
@@ -150,7 +150,7 @@ int VFPRINTF(HANDLE handle, const std::string& fmt)
 
 	if( !is_console(handle) && stdout_with_ansisequence )
 	{
-		WriteFile(handle,fmt.c_str(),fmt.length(),&written,0);
+		WriteFile(handle,fmt.c_str(),(DWORD)fmt.length(),&written,0);
 		return 0;
 	}
 	// start with processing
